# api-documentation/swagger/api.yaml
...
openapi: 3.0.3
info:
  title: Warehouse api
  description: Warehouse distribution api
  contact:
    name: Lum Berisha
    email: lum.berisha@frakton.com
  version: '1.0'
servers:
  - url: http://localhost:8080
    # description: Inferred Url
...
paths:
  /api/address:
    get:
      tags:
        - Address
      summary: getAddress
      operationId: getAddressUsingGET
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AddressResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Address
      summary: createAddress
      operationId: createAddressUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequestDto'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AddressResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/address/{id}:
    get:
      tags:
        - Address
      summary: getAddressById
      operationId: getAddressByIdUsingGET
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AddressResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - Address
      summary: updateAddress
      operationId: updateAddressUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequestDto'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AddressResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - Address
      summary: deleteAddress
      operationId: deleteAddressUsingDELETE
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AddressResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/employee/address/{employeeId}/{addressId}:
    post:
      tags:
        - Employee
      summary: setAddress
      operationId: setAddressUsingPOST
      parameters:
        - name: employeeId
          in: path
          description: employeeId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
        - name: addressId
          in: path
          description: addressId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/EmployeeResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/order/address/{orderId}:
    get:
      tags:
        - Order
      summary: getClientAddressFromOrder
      operationId: getClientAddressFromOrderUsingGET
      parameters:
        - name: orderId
          in: path
          description: orderId
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                title: BaseResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AddressResponseDto'
                  code:
                    type: string
                    enum:
                      - ACCEPTED
                      - ALREADY_REPORTED
                      - BAD_GATEWAY
                      - BAD_REQUEST
                      - BANDWIDTH_LIMIT_EXCEEDED
                      - CHECKPOINT
                      - CONFLICT
                      - CONTINUE
                      - CREATED
                      - DESTINATION_LOCKED
                      - EXPECTATION_FAILED
                      - FAILED_DEPENDENCY
                      - FORBIDDEN
                      - FOUND
                      - GATEWAY_TIMEOUT
                      - GONE
                      - HTTP_VERSION_NOT_SUPPORTED
                      - IM_USED
                      - INSUFFICIENT_SPACE_ON_RESOURCE
                      - INSUFFICIENT_STORAGE
                      - INTERNAL_SERVER_ERROR
                      - I_AM_A_TEAPOT
                      - LENGTH_REQUIRED
                      - LOCKED
                      - LOOP_DETECTED
                      - METHOD_FAILURE
                      - METHOD_NOT_ALLOWED
                      - MOVED_PERMANENTLY
                      - MOVED_TEMPORARILY
                      - MULTIPLE_CHOICES
                      - MULTI_STATUS
                      - NETWORK_AUTHENTICATION_REQUIRED
                      - NON_AUTHORITATIVE_INFORMATION
                      - NOT_ACCEPTABLE
                      - NOT_EXTENDED
                      - NOT_FOUND
                      - NOT_IMPLEMENTED
                      - NOT_MODIFIED
                      - NO_CONTENT
                      - OK
                      - PARTIAL_CONTENT
                      - PAYLOAD_TOO_LARGE
                      - PAYMENT_REQUIRED
                      - PERMANENT_REDIRECT
                      - PRECONDITION_FAILED
                      - PRECONDITION_REQUIRED
                      - PROCESSING
                      - PROXY_AUTHENTICATION_REQUIRED
                      - REQUESTED_RANGE_NOT_SATISFIABLE
                      - REQUEST_ENTITY_TOO_LARGE
                      - REQUEST_HEADER_FIELDS_TOO_LARGE
                      - REQUEST_TIMEOUT
                      - REQUEST_URI_TOO_LONG
                      - RESET_CONTENT
                      - SEE_OTHER
                      - SERVICE_UNAVAILABLE
                      - SWITCHING_PROTOCOLS
                      - TEMPORARY_REDIRECT
                      - TOO_EARLY
                      - TOO_MANY_REQUESTS
                      - UNAUTHORIZED
                      - UNAVAILABLE_FOR_LEGAL_REASONS
                      - UNPROCESSABLE_ENTITY
                      - UNSUPPORTED_MEDIA_TYPE
                      - UPGRADE_REQUIRED
                      - URI_TOO_LONG
                      - USE_PROXY
                      - VARIANT_ALSO_NEGOTIATES
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
components:
  schemas:
    AddressRequestDto:
      title: AddressRequestDto
      required:
        - city
        - region
        - street
      type: object
      properties:
        city:
          type: string
        postalCode:
          type: integer
          format: int32
        region:
          type: string
          enum:
            - GJAKOVE
            - PEJE
            - PRISHTINE
            - PRIZREN
        street:
          type: string
...
    EmployeeResponseDto:
      title: EmployeeResponseDto
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponseDto'
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        address:
          $ref: '#/components/schemas/AddressResponseDto'
        vehicle:
          $ref: '#/components/schemas/VehicleResponseDto'
    VehicleResponseDto:
      title: VehicleResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        plateNumber:
          type: string
        personCapacity:
          type: integer
          format: int32
        vehicleType:
          type: string
          enum:
            - PICKUP
            - VAN
            - TRACK
    AddressResponseDto:
      title: AddressResponseDto
      type: object
      properties:
        id:
          type: string
          format: uuid
        region:
          type: string
          enum:
            - GJAKOVE
            - PEJE
            - PRISHTINE
            - PRIZREN
        street:
          type: string
        city:
          type: string
        postalCode:
          type: integer
          format: int32
...
    UserResponseDto:
      title: UserResponseDto
      properties:
        id:
          type: string
          format: uuid
        userName:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - CLIENT
            - OFFICE_WORKER
            - DELIVERY
            - WAREHOUSE_WORKER
...
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header